+++ src/host/mod.rs
@@ -208,32 +208,7 @@ pub(crate) mod error_emit;
         )
     ),
 ))]
-pub(crate) use error_emit::emit_error;
-#[cfg(any(
-    target_vendor = "apple",
-    all(target_os = "android", feature = "realtime"),
-    all(
-        feature = "jack",
-        any(
-            target_os = "linux",
-            target_os = "dragonfly",
-            target_os = "freebsd",
-            target_os = "netbsd",
-            target_os = "macos",
-            target_os = "windows",
-        )
-    ),
-    all(
-        feature = "pipewire",
-        any(
-            target_os = "linux",
-            target_os = "dragonfly",
-            target_os = "freebsd",
-            target_os = "netbsd",
-        )
-    ),
-))]
-pub(crate) use error_emit::try_emit_error;
+pub(crate) use error_emit::{emit_error, try_emit_error};
 
 /// Convert a frame count at a given sample rate to a [`std::time::Duration`].
 #[cfg(any(